Add PR template to the repo
We want to make sure people format their code as well as reference the
issue that's going to be closed as well so we don't have to close the
issue manually! This PR adds a template for PRs that includes these
points as a checklist to ensure future contributors know what they need
to do!