fix(test): Remove lock/update wasm-bindgen
This does a few things, remove Cargo.lock from tracking to make sure
when compiled the newest version is used in a fresh repo. Testing on
one's own will require running an 'rm tests/fixtures/*/Cargo.lock' if
wanting to bump to the newest version of wasm-bindgen. This also sets
all tests to use the latest version of wasm-bindgen in 0.2. This shows
that we support that version (0.2.x) but bumping it to 0.3 means we've
opted for that kind of change, but for now it's not been bumped to that.