feat(log): Log information as program runs
This commit does a few things:

- Sets up the ability to log information based off log level
- Figures out where to log the file
- Starts logging information in the program

As this is a first pass the logging is good enough to know where things
went wrong, but we can expand on this in the future to log many many
things. We just now have an initial implementation. The log is written
out to a file that can be read by the user if they pass in -v, -vv, etc.
where each v is an extra level of logging.