Remove `branch: master` from AppVeyor deploy
It looks like Windows releases aren't getting build for wasm-pack tags
unfortunately and taking a look at the [last logs][logs] it looks like the
reason is that the `master` branch is required. I'm not actually entirely sure
what's required here but by removing this it's closer to `wasm-bindgen`'s
configuration which I think is working!

[logs]: https://ci.appveyor.com/project/ashleygwilliams/wasm-pack-071k0/build/1.0.72