Use a global cache for all downloaded binaries
This commit switches wasm-pack to using a global cache for all download
binaries, living typically in a user's home directory. The intention
here is to aovid creating a `bin` folder in all wasm-pack projects and
additionally share downloads between projects to ensure that you're
downloading a minimal number of binaries from the network.

Along the way the downloading code was restructured to support a global
cache, but everything should largely be as it was before!

Closes #292