Flag gh-pages docs as unpublished
This commit adds a header dialog to the gh-pages mdbook documentation of
this repository to indicate that it represents unpublished documentation
that may not reflect the currently published version of `wasm-pack`. The
header redirects users to https://rustwasm.github.io/docs/wasm-pack/
which has the documentation for the latest published version of
`wasm-pack`.