Search for tools in download_prebuilt
Unlike other tools, wasm-opt doesn't have an option to `cargo install`, so the `which` check in `download_prebuilt_or_cargo_install` wasn't applicable.

It makes more sense to move such check to the more generic function `download_prebuilt`, in which case it's going to work for all tools in the same way.

Additionally, if the found version doesn't match the expected one, but user has specified `--mode=no-install`, try to use the tool anyway - most likely user knows what they're doing, and the version is backward-compatible, but if not, worst case the tool will error out on invalid params.

Fixes #869.