Update docs for the Rust-generated wasm function
The documentation is out of date. It says that there should be a `run()` function, but really a `main_js()` function should be there.
However, the `rust-webpack-template` doesn't add the `then()` statement to the import in the template.
I have created [a pull request](https://github.com/rustwasm/rust-webpack-template/pull/172) to add this to the template, and I have updated the docs for this newer version here.