cargo update
This lockfile seems very old and in a different syntax.  I need to
update parking_lot and cargo wants to update everything.

In my environment, the tests do the same thing before and after:
everything passes, except for
    test::it_can_find_a_webdriver_on_path
    test::it_can_run_browser_tests
which seem to be suffering some kind of problem with either my
installation or with an upstream service they use.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>