Centralise initial cargo command building
Replace all open-coded instances of Command::new("cargo") with a call
to global_opts().  This will allow the global options to influence how
we run cargo.

No functional change at this stage.

I did a git grep for '"cargo"' to find all the occurrences.  One
open-coded instance remains, in tests/all/utils/fixture.rs.  This is
not part of wasm-pack; rather, it is test machinery, so it should not
try to use wasm-pack's global_opts().

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>