Updated package.json structs and populated new main field
Added `main` field to  NoModulesPackage and ESModulesPackage in the generated package.json.

Also cloned the entry filename during struct instantiation for when the package.json is generated. Cloning might not be optimal, it is probably possible to borrow the String and generate the filename with that, but I think clones here should be fine.